requirements.txt
ollama
piper-tts
pygame